Added to the function of Update_Time - AN
